cleanup: use more appropriately ConfigurationProperties making a DTO for TTL heartbeating
